requirements.txt
Flask==3.0.0
mysqlclient==2.2.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
openai==1.26.0
python-dotenv==1.0.1